backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
werkzeug==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
SQLAlchemy==2.0.27